// src/app/basicinfo/person/models/person-edit.ts
export interface PersonEdit {
    id: number;
    lastName: string;
    firstName: string;
    nationalCode: string;
    fatherName: string;
    birthDate: Date;
    educationDegreeId: number;
    organizationName: string;
    organizationId: number;
    employmentDate: Date;
    rankId: number;
    studyFieldId: number;
    personnelCode: string;
}